+++ src/cosmology/api/compat/__init__.py
@@ -1,9 +1,9 @@
 """The Cosmology API standard for compatability wrappers."""
 
-from cosmology.api.compat._core import CosmologyWrapperAPI
-from cosmology.api.compat._standard import StandardCosmologyWrapperAPI
+from cosmology.api.compat._core import CosmologyWrapper
+from cosmology.api.compat._standard import StandardCosmologyWrapper
 
 __all__ = [
-    "CosmologyWrapperAPI",
-    "StandardCosmologyWrapperAPI",
+    "CosmologyWrapper",
+    "StandardCosmologyWrapper",
 ]